-zオプションで場合によっては強くぼやけた画像になってしまう対策（roundfに変更）

+++ shadow/main.m
@@ -144,18 +144,18 @@ NSImage* trimImageByRect(NSImage *image, NSRect trimRect)
 NSSize resizeWidth(NSSize aSize, float pxRate)
 {
     if (pxRate > BBKRateLimit) {
-        return NSMakeSize(pxRate, round(aSize.height * pxRate / aSize.width));
+        return NSMakeSize(pxRate, roundf(aSize.height * pxRate / aSize.width));
     }else{
-        return NSMakeSize(round(aSize.width * pxRate), round(aSize.height * pxRate));
+        return NSMakeSize(roundf(aSize.width * pxRate), roundf(aSize.height * pxRate));
     }
 }
 
 NSSize resizeHeight(NSSize aSize, float pxRate)
 {
     if (pxRate > BBKRateLimit) {
-        return NSMakeSize(round(aSize.width * pxRate / aSize.height), pxRate);
+        return NSMakeSize(roundf(aSize.width * pxRate / aSize.height), pxRate);
     }else{
-        return NSMakeSize(round(aSize.width * pxRate), round(aSize.height * pxRate));
+        return NSMakeSize(roundf(aSize.width * pxRate), roundf(aSize.height * pxRate));
     }
 }
